Finish migration to standalone app
