+++ client/src/app/modules/core/core.component.ts
@@ -11,12 +11,27 @@ import {
 } from '../../ngrx/selectors/instance-settings.selectors';
 import { take } from 'rxjs/operators';
 import { NavigationService } from '../../services/core/navigation.service';
+import { BackgroundImageComponent } from './background-image/background-image.component';
+import { AppLevelAlertsComponent } from './app-level-alerts/app-level-alerts.component';
+import { MenuComponent } from './menu/menu.component';
+import { RouterOutlet } from '@angular/router';
+import { FooterComponent } from './footer/footer.component';
+import { RefreshLoginModalComponent } from './refresh-login-modal/refresh-login-modal.component';
+import { Toast } from 'primeng/toast';
 
 @Component({
   selector: 'lc-root',
   templateUrl: './core.component.html',
   styleUrls: ['./core.component.scss'],
-  standalone: false,
+  imports: [
+    BackgroundImageComponent,
+    AppLevelAlertsComponent,
+    MenuComponent,
+    RouterOutlet,
+    FooterComponent,
+    RefreshLoginModalComponent,
+    Toast,
+  ],
 })
 export class CoreComponent implements OnInit {
   constructor(
